move Message and Topic to forum folder
