--- src/components/ForumPageComponents/Topic/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Topic';
