+++ src/components/forum/Topic/index.ts
@@ -1,0 +1,1 @@
+export * from './Topic';
