+++ src/components/forum/index.ts
@@ -1,0 +1,2 @@
+export * from './Message';
+export * from './Topic';
