+++ src/components/index.ts
@@ -1,5 +1,6 @@
 // такой индекс-файл нужен для того чтобы можно было импортировать
 // компонент просто указав путь до папки components
 
-export * from './common';
 export * from './auth';
+export * from './common';
+export * from './forum';
